Refactor Vibrate Control API
Refactor Vibrate Control API as Singleton (for using as Manager), now can use everywhere where XR Origin object is activated and assigned its controllers
