+++ Assets/00. Workspace/03. Park/VibrateTest/VibrateTestUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8936295052a487041a27cebd41207905
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
