+++ Assets/02. Scripts/Vibration.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8e11508493abb7d4eae2fa5ac8b848d3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
