+++ Assets/03. Prefabs/Managers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a045dea6ee4267a41924c6c8c6dc2f7d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
