+++ Assets/03. Prefabs/Managers/Vibrate Controller.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c334a348ab2e4064f825b7680dd38ab0
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
